requirements.txt
# ============================================================================
# PYTHON PACKAGE REQUIREMENTS:
# ============================================================================
# REQUIRES: Python >= 3.7
# DESCRIPTION:
#    pip install -r <THIS_FILE>
#
# SPECIFIC DEPENDENCIES:
#   behave  -- BDD test framework, cucumber-clone for Python.
#   selenium -- To drive the browser tests
#   PyHamcrest -- For assertions: assert_that(expected, equal_to(actual))
#   PyYAML -- For working with yaml files
#   Allure-Behave -- For HTML report
#   Faker -- For working with faker data
# ============================================================================

PyYAML >= 3.11
behave >= 1.2.5
pyodbc
parse >= 1.8.4
parse-type >= 0.4.2
selenium >= 2.53.1
#argparse >= 1.4.0
PyHamcrest >= 1.8
allure-behave
Faker == 0.9.0
